Refatorando arquivo de testes Placa de Carro
