+++ brutils/pis.py
@@ -2,6 +2,8 @@ from random import randint
 
 
 WEIGHTS = [3, 2, 9, 8, 7, 6, 5, 4, 3, 2]
+# FORMATTING
+############
 
 
 # OPERATIONS
